refactor(issue-page): split issue page into client and wrapper components
- Extracted client-side logic to issue-page-client.tsx
- Added issue-page-wrapper.tsx for server-side data fetching and validation
- Updated main issue page to use wrapper component
- Added IssueNotFound component for handling missing issues
- Improved error handling in IssueDetails component
