+++ app/[slug]/issue/[id]/not-found.tsx
@@ -1,0 +1,11 @@
+import { IssueNotFound } from '@/components/errors/issue-not-found'
+
+export default function NotFound({
+  params
+}: {
+  params?: { slug?: string }
+}) {
+  const workspaceSlug = params?.slug || 'workspace'
+  
+  return <IssueNotFound workspaceSlug={workspaceSlug} />
+}
